Moved winorlose into a new file
