+++ gameComponents/gameVars.py
@@ -1,6 +1,6 @@
 choices = ["rock", "paper", "scissors"]
-player_lives = 3
-ai_lives = 3
-total_lives = 3
+total_lives = 1
+player_lives = total_lives
+ai_lives = total_lives
 # Booleans are an on or off switch, 1 or 0, true or false
 player_choice = False
